requirements.txt
asgiref==3.8.1
Django==5.0.7
django-bootstrap5==24.2
django-filter==24.2
django-phonenumber-field==8.0.0
factory-boy==3.3.0
Faker==26.0.0
phonenumbers==8.13.40
pillow==10.4.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
